feat: reliable decoding of signals
