+++ signal_layout_test.go
@@ -1,147 +1,386 @@
 package acmelib
 
-// import (
-// 	"testing"
+import (
+	"testing"
 
-// 	"github.com/stretchr/testify/assert"
-// )
+	"github.com/stretchr/testify/assert"
+)
 
-// func Test_SignalLayout_Unpack(t *testing.T) {
-// 	assert := assert.New(t)
+func Test_SignalLayout_insert(t *testing.T) {
+	assert := assert.New(t)
 
-// 	// Define all signal types
-// 	uint12, err := NewIntegerSignalType("uint12_t", 12, false)
-// 	assert.NoError(err)
-// 	int4, err := NewIntegerSignalType("int4_t", 4, true)
-// 	assert.NoError(err)
-// 	flag := NewFlagSignalType("flag_t")
-// 	int20, err := NewIntegerSignalType("int20_t", 20, true)
-// 	assert.NoError(err)
+	sl := newSL(8)
 
-// 	// Define the signal enum
-// 	enum := NewSignalEnum("enum_e")
-// 	enum.SetMinSize(4)
-// 	enum.AddValue(NewSignalEnumValue("OK", 1))
+	// All signals are 8 bits long
+	sigType, err := NewIntegerSignalType("8_bits", 8, false)
+	assert.NoError(err)
 
-// 	// Define the little endian signals
-// 	sig0le, err := NewStandardSignal("sig_0_le", uint12)
-// 	assert.NoError(err)
-// 	sig1le, err := NewStandardSignal("sig_1_le", int4)
-// 	assert.NoError(err)
-// 	sig2le, err := NewStandardSignal("sig_2_le", int20)
-// 	assert.NoError(err)
-// 	sig3le, err := NewEnumSignal("sig_3_le", enum)
-// 	assert.NoError(err)
-// 	sig4le, err := NewStandardSignal("sig_4_le", flag)
-// 	assert.NoError(err)
+	sig0, err := NewStandardSignal("sig_0", sigType)
+	assert.NoError(err)
 
-// 	// Define the little endian message
-// 	msgle := NewMessage("msg_le", 1, 8)
-// 	assert.NoError(msgle.AppendSignal(sig0le))
-// 	assert.NoError(msgle.InsertSignal(sig1le, 28))
-// 	assert.NoError(msgle.InsertSignal(sig2le, 34))
-// 	assert.NoError(msgle.InsertSignal(sig3le, 58))
-// 	assert.NoError(msgle.InsertSignal(sig4le, 63))
+	// Should not return an error
+	assert.NoError(sl.verifyAndInsert(sig0, 0))
 
-// 	// Get the signal layout for the little endian message
-// 	slle := msgle.SignalLayout()
+	sig1, err := NewStandardSignal("sig_1", sigType)
+	assert.NoError(err)
 
-// 	// Define the big endian signals
-// 	sig0be, err := NewStandardSignal("sig_0_be", uint12)
-// 	assert.NoError(err)
-// 	sig1be, err := NewStandardSignal("sig_1_be", int4)
-// 	assert.NoError(err)
-// 	sig2be, err := NewStandardSignal("sig_2_be", int20)
-// 	assert.NoError(err)
-// 	sig3be, err := NewEnumSignal("sig_3_be", enum)
-// 	assert.NoError(err)
-// 	sig4be, err := NewStandardSignal("sig_4_be", flag)
-// 	assert.NoError(err)
+	// Should return an error
+	assert.Error(sl.verifyAndInsert(sig1, 0))
 
-// 	// Define the big endian message
-// 	msgbe := NewMessage("msg_be", 1, 8)
-// 	assert.NoError(msgbe.AppendSignal(sig0be))
-// 	assert.NoError(msgbe.InsertSignal(sig1be, 28))
-// 	assert.NoError(msgbe.InsertSignal(sig2be, 34))
-// 	assert.NoError(msgbe.InsertSignal(sig3be, 58))
-// 	assert.NoError(msgbe.InsertSignal(sig4be, 63))
-// 	msgbe.SetByteOrder(MessageByteOrderBigEndian)
+	// Should not return an error
+	assert.NoError(sl.verifyAndInsert(sig1, 16))
 
-// 	// Get the signal layout for the big endian message
-// 	slbe := msgbe.SignalLayout()
+	sig2, err := NewStandardSignal("sig_2", sigType)
+	assert.NoError(err)
 
-// 	// Test the value of unpacked signals
-// 	expectedValues := []any{uint64(2049), int64(-7), int64(-524287), "OK", true}
+	// Should return an error in each case
+	assert.Error(sl.verifyAndInsert(sig2, -1))
+	assert.Error(sl.verifyAndInsert(sig2, 63))
+	assert.Error(sl.verifyAndInsert(sig2, 57))
+	assert.Error(sl.verifyAndInsert(sig2, 7))
+	assert.Error(sl.verifyAndInsert(sig2, 23))
 
-// 	data0 := []byte{
-// 		0b00000001,
-// 		0b00001000,
-// 		0b00000000,
-// 		0b10010000,
-// 		0b00000100,
-// 		0b00000000,
-// 		0b00100000,
-// 		0b10000100,
-// 	}
+	// Should not return an error
+	assert.NoError(sl.verifyAndInsert(sig2, 8))
 
-// 	for idx, dec := range slle.Decode(data0) {
-// 		assert.Equal(expectedValues[idx], dec.Value)
-// 	}
+	// Should have 3 signals and filters
+	assert.Len(sl.Signals(), 3)
+	assert.Len(sl.Filters(), 3)
+}
 
-// 	data1 := []byte{
-// 		0b10000000,
-// 		0b00010000,
-// 		0b00000000,
-// 		0b10010000,
-// 		0b00100000,
-// 		0b00000000,
-// 		0b00000100,
-// 		0b10000100,
-// 	}
+func Test_SignalLayout_delete(t *testing.T) {
+	assert := assert.New(t)
 
-// 	for idx, dec := range slbe.Decode(data1) {
-// 		assert.Equal(expectedValues[idx], dec.Value)
-// 	}
+	sl := newSL(8)
 
-// 	// Testing multiplexed signals
+	// All signals are 8 bits long
+	sigType, err := NewIntegerSignalType("8_bits", 8, false)
+	assert.NoError(err)
 
-// 	// sigMux0Little, err := NewStandardSignal("sig_mux_0_le", int4)
-// 	// assert.NoError(err)
-// 	// sigMux1Little, err := NewStandardSignal("sig_mux_1_le", int20)
-// 	// assert.NoError(err)
-// 	// sigMux2Little, err := NewEnumSignal("sig_mux_2_le", enum)
-// 	// assert.NoError(err)
-// 	// sigMux3Little, err := NewStandardSignal("sig_mux_3_le", flag)
-// 	// assert.NoError(err)
+	sig0, err := NewStandardSignal("sig_0", sigType)
+	assert.NoError(err)
+	sig1, err := NewStandardSignal("sig_1", sigType)
+	assert.NoError(err)
+	sig2, err := NewStandardSignal("sig_2", sigType)
+	assert.NoError(err)
 
-// 	// sigMuxorLittle, err := NewMultiplexerSignal("sig_muxor_le", 4096, 52)
-// 	// assert.NoError(err)
-// 	// assert.Equal(12, sigMuxorLittle.GetGroupCountSize())
+	assert.NoError(sl.verifyAndInsert(sig0, 0))
+	assert.NoError(sl.verifyAndInsert(sig1, 8))
+	assert.NoError(sl.verifyAndInsert(sig2, 16))
 
-// 	// assert.NoError(sigMuxorLittle.InsertSignal(sigMux0Little, 16, 0))
-// 	// assert.NoError(sigMuxorLittle.InsertSignal(sigMux1Little, 22, 1))
-// 	// assert.NoError(sigMuxorLittle.InsertSignal(sigMux2Little, 46, 2))
-// 	// assert.NoError(sigMuxorLittle.InsertSignal(sigMux3Little, 51, 2049))
+	// Should have 3 signals
+	assert.Len(sl.Signals(), 3)
 
-// 	// msgMuxLittle := NewMessage("msg_mux_le", 1, 8)
-// 	// assert.NoError(msgMuxLittle.AppendSignal(sigMuxorLittle))
+	// Should delete signal 1
+	sl.delete(sig1)
 
-// 	// slMuxLittle := msgMuxLittle.SignalLayout()
+	// Should have 2 signals and filters
+	assert.Len(sl.Signals(), 2)
+	assert.Len(sl.Filters(), 2)
+}
 
-// 	// log.Print(slMuxLittle)
+func Test_SignalLayout_clear(t *testing.T) {
+	assert := assert.New(t)
 
-// 	// dataMuxLittle := []byte{
-// 	// 	0b00000001,
-// 	// 	0b00001000,
-// 	// 	0b00000000,
-// 	// 	0b10010000,
-// 	// 	0b00000100,
-// 	// 	0b00000000,
-// 	// 	0b00100000,
-// 	// 	0b10000100,
-// 	// }
+	sl := newSL(8)
 
-// 	// for _, dec := range slMuxLittle.Decode(dataMuxLittle) {
-// 	// 	log.Print(dec)
-// 	// }
-// }
+	// All signals are 8 bits long
+	sigType, err := NewIntegerSignalType("8_bits", 8, false)
+	assert.NoError(err)
+
+	sig0, err := NewStandardSignal("sig_0", sigType)
+	assert.NoError(err)
+	sig1, err := NewStandardSignal("sig_1", sigType)
+	assert.NoError(err)
+	sig2, err := NewStandardSignal("sig_2", sigType)
+	assert.NoError(err)
+
+	assert.NoError(sl.verifyAndInsert(sig0, 0))
+	assert.NoError(sl.verifyAndInsert(sig1, 8))
+	assert.NoError(sl.verifyAndInsert(sig2, 16))
+
+	// Should have 3 signals
+	assert.Len(sl.Signals(), 3)
+
+	// Should have 0 signals and filters
+	sl.clear()
+	assert.Len(sl.Signals(), 0)
+	assert.Len(sl.Filters(), 0)
+}
+
+func Test_SignalLayout_updateStartPos(t *testing.T) {
+	assert := assert.New(t)
+
+	sl := newSL(8)
+
+	// All signals are 8 bits long
+	sigType, err := NewIntegerSignalType("8_bits", 8, false)
+	assert.NoError(err)
+
+	sig0, err := NewStandardSignal("sig_0", sigType)
+	assert.NoError(err)
+	sig1, err := NewStandardSignal("sig_1", sigType)
+	assert.NoError(err)
+
+	assert.NoError(sl.verifyAndInsert(sig0, 0))
+	assert.NoError(sl.verifyAndInsert(sig1, 8))
+
+	// Should not return an error
+	assert.NoError(sl.verifyAndUpdateStartPos(sig1, 16))
+
+	// Should return an error
+	assert.Error(sl.verifyAndUpdateStartPos(sig1, -1))
+	assert.Error(sl.verifyAndUpdateStartPos(sig1, 63))
+	assert.Error(sl.verifyAndUpdateStartPos(sig1, 57))
+	assert.Error(sl.verifyAndUpdateStartPos(sig1, 7))
+
+	// Should not return an error
+	assert.NoError(sl.verifyAndUpdateStartPos(sig0, 8))
+
+	// Should return an error
+	assert.Error(sl.verifyAndUpdateStartPos(sig0, 15))
+
+	// Should have 2 signals and filters
+	assert.Len(sl.Signals(), 2)
+	assert.Len(sl.Filters(), 2)
+}
+
+func Test_SignalLayout_updateSize(t *testing.T) {
+	assert := assert.New(t)
+
+	sl := newSL(8)
+
+	// All signals are 8 bits long at start
+	sigType, err := NewIntegerSignalType("8_bits", 8, false)
+	assert.NoError(err)
+
+	sig0, err := NewStandardSignal("sig_0", sigType)
+	assert.NoError(err)
+	sig1, err := NewStandardSignal("sig_1", sigType)
+	assert.NoError(err)
+
+	assert.NoError(sl.verifyAndInsert(sig0, 0))
+	assert.NoError(sl.verifyAndInsert(sig1, 8))
+
+	// Should not return an error
+	assert.NoError(sl.verifyAndUpdateSize(sig1, 32))
+
+	assert.Equal(32, sig1.Size())
+
+	// Should return an error
+	assert.Error(sl.verifyAndUpdateSize(sig0, -1))
+	assert.Error(sl.verifyAndUpdateSize(sig0, 9))
+	assert.Error(sl.verifyAndUpdateSize(sig0, 65))
+
+	// Should have 2 signals and 5 filters
+	assert.Len(sl.Signals(), 2)
+	assert.Len(sl.Filters(), 5)
+}
+
+func Test_SignalLayout_Compact(t *testing.T) {
+	assert := assert.New(t)
+
+	tdBasicMsg := initBasicMessage(assert)
+
+	msgBasic := tdBasicMsg.message
+	assert.NoError(msgBasic.InsertSignal(dummySignal, 56))
+
+	tdBasicMsg.layout.Compact()
+	assert.Equal(48, dummySignal.StartPos())
+
+	assert.NoError(msgBasic.DeleteSignal(dummySignal.EntityID()))
+}
+
+func Test_SignalLayout_Decode(t *testing.T) {
+	assert := assert.New(t)
+
+	tdBasicMsg := initBasicMessage(assert)
+	layout := tdBasicMsg.layout
+
+	// Every raw value will be 11xxx1 where x are all 0s depending on the size of the signal
+	data := []byte{
+		0b00000001,
+		0b00011100,
+		0b11000000,
+		0b00000001,
+		0b00011100,
+		0b11000000,
+		0b00000000,
+		0b00000000,
+	}
+
+	expectedRaw := []any{uint64(3073), uint64(3073), uint64(3073), uint64(3073)}
+	decodings := layout.Decode(data)
+	assert.Len(decodings, 4)
+	for idx, dec := range decodings {
+		assert.Equal(expectedRaw[idx], dec.RawValue)
+	}
+
+	tdTypedMsg := initTypedMessage(assert)
+	layout = tdTypedMsg.layout
+
+	data = []byte{
+		0b00000001,
+		0b11000001,
+		0b11000001,
+		0b00000001,
+		0b11000000,
+		0b00000001,
+		0b11000000,
+	}
+
+	expectedDec := []SignalDecoding{
+		{ValueType: SignalValueTypeFlag, RawValue: 1, Value: true},
+		{ValueType: SignalValueTypeUint, RawValue: 193, Value: uint64(193)},
+		{ValueType: SignalValueTypeInt, RawValue: ^uint64(63 - 1), Value: int64(-63)},
+		{ValueType: SignalValueTypeFloat, RawValue: 49153, Value: float64(24677), Unit: "V"},              // scale: 0.5; offset: 100.5
+		{ValueType: SignalValueTypeFloat, RawValue: ^uint64(16383 - 1), Value: float64(-8091), Unit: "V"}, // scale: 0.5; offset: 100.5
+	}
+
+	decodings = layout.Decode(data)
+	assert.Len(decodings, 5)
+	for idx, dec := range decodings {
+		assert.Equal(expectedDec[idx].ValueType, dec.ValueType)
+		assert.Equal(expectedDec[idx].RawValue, dec.RawValue)
+		assert.Equal(expectedDec[idx].Value, dec.Value)
+		assert.Equal(expectedDec[idx].Unit, dec.Unit)
+	}
+
+	tdEnumMsg := initEnumMessage(assert)
+	layout = tdEnumMsg.layout
+
+	data = []byte{
+		0b00000001,
+		0b00000010,
+		0b01111111,
+		0b00000000,
+	}
+
+	expectedDec = []SignalDecoding{
+		{ValueType: SignalValueTypeEnum, RawValue: 1, Value: "enum_value_1"},
+		{ValueType: SignalValueTypeEnum, RawValue: 2, Value: "enum_value_2"},
+		{ValueType: SignalValueTypeEnum, RawValue: 127, Value: "enum_value_127"},
+	}
+
+	decodings = layout.Decode(data)
+	assert.Len(decodings, 3)
+	for idx, dec := range decodings {
+		assert.Equal(expectedDec[idx].ValueType, dec.ValueType)
+		assert.Equal(expectedDec[idx].RawValue, dec.RawValue)
+		assert.Equal(expectedDec[idx].Value, dec.Value)
+	}
+
+	tdMuxMsg := initMuxMessage(assert)
+	layout = tdMuxMsg.layout
+
+	data = []byte{
+		0b00000001, // selecting layout 1 from top
+		0b11111111, // selecting layout 255 from top inner
+		0b11000001, // top inner in 255
+		0b00000001,
+		0b11000000,
+		0b11000001, // bottom in 0,2
+		0b11000001, // bottom in 0
+		0b00000000, // selecting layout 0 from bottom
+	}
+
+	expectedDec = []SignalDecoding{
+		{
+			Signal:    tdMuxMsg.layers.top.layer.muxor,
+			ValueType: SignalValueTypeUint,
+			RawValue:  1,
+			Value:     uint64(1),
+		},
+		{
+			Signal:    tdMuxMsg.layers.top.inner.layer.muxor,
+			ValueType: SignalValueTypeUint,
+			RawValue:  255,
+			Value:     uint64(255),
+		},
+		{
+			Signal:    tdMuxMsg.layers.top.inner.signals.in255,
+			ValueType: SignalValueTypeUint,
+			RawValue:  193,
+			Value:     uint64(193),
+		},
+		{
+			Signal:    tdMuxMsg.signals.base,
+			ValueType: SignalValueTypeUint,
+			RawValue:  49153,
+			Value:     uint64(49153),
+		},
+		{
+			Signal:    tdMuxMsg.layers.bottom.layer.muxor,
+			ValueType: SignalValueTypeUint,
+			RawValue:  0,
+			Value:     uint64(0),
+		},
+		{
+			Signal:    tdMuxMsg.layers.bottom.signals.in02,
+			ValueType: SignalValueTypeUint,
+			RawValue:  193,
+			Value:     uint64(193),
+		},
+		{
+			Signal:    tdMuxMsg.layers.bottom.signals.in0,
+			ValueType: SignalValueTypeUint,
+			RawValue:  193,
+			Value:     uint64(193),
+		},
+	}
+
+	decodings = layout.Decode(data)
+	assert.Len(decodings, 7)
+	for idx, dec := range decodings {
+		assert.Equal(expectedDec[idx].Signal.Name(), dec.Signal.Name())
+		assert.Equal(expectedDec[idx].ValueType, dec.ValueType)
+		assert.Equal(expectedDec[idx].RawValue, dec.RawValue)
+		assert.Equal(expectedDec[idx].Value, dec.Value)
+	}
+}
+
+func Benchmark_SignalLayout_Decode(b *testing.B) {
+	assert := assert.New(b)
+
+	tdEnumMsg := initEnumMessage(assert)
+	layout := tdEnumMsg.layout
+
+	data := []byte{
+		0b00000001,
+		0b00000010,
+		0b01111111,
+		0b00000000,
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for b.Loop() {
+		layout.Decode(data)
+	}
+}
+
+func Benchmark_SignalLayout_DecodeMultiplexed(b *testing.B) {
+	assert := assert.New(b)
+
+	tdMuxMsg := initMuxMessage(assert)
+	layout := tdMuxMsg.layout
+
+	data := []byte{
+		0b00000001, // selecting layout 1 from top
+		0b11111111, // selecting layout 255 from top inner
+		0b11000001, // top inner in 255
+		0b00000001,
+		0b11000000,
+		0b11000001, // bottom in 0,2
+		0b11000001, // bottom in 0
+		0b00000000, // selecting layout 0 from bottom
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for b.Loop() {
+		layout.Decode(data)
+	}
+}
